% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data.R
\name{build_data}
\alias{build_data}
\title{Run all scripts to build data}
\usage{
build_data(dir = fs::path("data-raw", "R.d"), verbose = TRUE)
}
\arguments{
\item{dir}{Directory to use for R.d files (default is \code{data-raw/R.d}).}

\item{verbose}{Should the files processed be printed as they are executed.}
}
\value{
Logical value (TRUE) invisibly
}
\description{
Run all scripts to build data
}
\examples{
\dontrun{
 # Source all scripts in data-raw/R.d:
 build_data()
 # Source all scripts in alternate directory:
 build_data(dir = "~/mydir/R.dx")
}
}
