% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{datatable}
\alias{datatable}
\title{Title}
\usage{
datatable(data, ...)
}
\arguments{
\item{data}{a data object (either a matrix or a data frame)}

\item{...}{Other options arguments (see \code{\link[DT:datatable]{DT::datatable()}}).}
}
\description{
Title
}
