% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvc_import.R
\name{dvc_import}
\alias{dvc_import}
\title{Import data to project}
\usage{
dvc_import(url, dest = NULL, import_dir = "data-raw/imports")
}
\arguments{
\item{import_dir}{}
}
\description{
Import data to project
}
