% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maker.R
\name{fig_maker}
\alias{fig_maker}
\title{Title}
\usage{
fig_maker(
  x,
  label = "@fig-fig1",
  filename = knitr::current_input(),
  output = c("png", "pdf")
)
}
\arguments{
\item{output}{}
}
\description{
Title
}
