% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maker.R
\name{tbl_maker}
\alias{tbl_maker}
\title{Title}
\usage{
tbl_maker(
  x,
  label = "@fig-fig1",
  filename = knitr::current_input(),
  output = c("word", "excel")
)
}
\arguments{
\item{filename}{}
}
\description{
Title
}
